Display ??? when we cannot infer the arguments to a defgraft.
This can happen because our parsing code assumes that a defgraft is
defined in the same namespace (and after) the pipe that it composes.

In order to do this more robustly, and allow arg inference to occur when
pipes are defined in other namespaces we will need to either

1) parse the code in the same order that clojure does (by following all
the requires transitively) so we can guarantee we have seen any imported
pipes.

2) Build a big table of all pipes on the classpath, then derive all the
   data for the grafts.